Renamed the files, made src-layout
