+++ src/sudoku_solver_task/rest_solve_sudoku.py
@@ -1,8 +1,10 @@
-from decision_sudoku import solve
+from src.sudoku_solver_task.cli_solve_sudoku import validate_sudoku
+
 
 from flask import Flask, request
 
-from solve_sudoku import validate_sudoku
+
+from src.sudoku_solver_task.sudoku_solver import solve
 
 
 app = Flask(__name__)
